# server/docs/swagger.yaml
basePath: /api
definitions:
...
  auth.RegisterData:
    properties:
      email:
        type: string
      farm:
        type: integer
      fullname:
        type: string
      password:
        type: string
      phone:
        type: string
      region:
        type: integer
      role:
        type: integer
    type: object
  cows.Children:
    properties:
      breed_name:
        description: Порода ребенка
        type: string
      child:
        $ref: '#/definitions/models.Cow'
      hoz_name:
        description: Название хозяйства, в котором ребенок
        type: string
    type: object
...
  models.CheckMilk:
    properties:
      checkDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата конрольной дойки
      dryMatter:
        description: сухой материал
        type: number
      fat:
        description: Параметр контрольной дойки, как я понимаю кол-во жира в молоке
        example: 1
        type: number
      id:
        description: ID контрольной дойки
        example: 1
        type: integer
      lactationId:
        description: ID лактации для которой выполнена контрольная дойка
        example: 1
        type: integer
      milk:
        description: Параметр контрольной дойки, как я понимаю кол-во молока
        example: 1
        type: number
      probeNumber:
        description: номер пробы
        type: integer
      protein:
        description: Параметр контрольной дойки, как я понимаю кол-во белка в молоке
        example: 1
        type: number
      somaticNucCount:
        description: количество соматических клеток
        type: number
    type: object
  models.Cow:
    properties:
      approved:
        description: Целое число, 0 - корова не подтверждена, 1 - корова подтверждена,
          -1 - корова отклонена
        example: 1
        type: integer
      birkingDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата перебирковки
      birthDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: День рождения
      birthHozId:
        description: ID хозяйства рождения
        type: integer
      birthMethod:
        description: 'способ зачатия: клон, эмбрион, искусственное осеменени, естественное
          осеменение'
        type: string
      breedId:
        description: ID породы коровы
        example: 1
        type: integer
      createdAt:
        description: Время создания коровы в базе данных
        example: "2007-01-01"
        type: string
      deathDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата смерти
      departDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: День отбытия из коровника
      exterior:
        $ref: '#/definitions/models.Exterior'
      farmGroupId:
        description: ID хозяйства, которому корова принадлежит
        example: 1
        type: integer
      farmID:
        description: ID фермы, которой корова принадлежит
        example: 1
        type: integer
      fatherSelecs:
        description: ID коровы отца коровы
        type: integer
      genetic:
        $ref: '#/definitions/models.Genetic'
      gradeHozId:
        description: оценка по хозяйству
        example: 1
        type: integer
      gradeRegionId:
        description: оценка по региону
        example: 1
        type: integer
      holding:
        $ref: '#/definitions/models.Farm'
      holdingID:
        description: ID холдинга, которому принадлежит корова
        type: integer
      id:
        description: ID коровы
        example: 1
        type: integer
      identificationNumber:
        description: он все-таки есть! это какой-то не российский номер коровы
        type: string
      inbrindingCoeffByFamily:
        description: Exterior                float64  `example:"3.14"` // Оценка экстерьера
          коровы, будет переделано в ID экстерьера коровы
        example: 3.14
        type: number
      inventoryNumber:
        description: Инвентарный номер коровы
        example: "1213321"
        type: string
      motherSelecs:
        description: ID коровы матери коровы
        type: integer
      name:
        description: Кличка коровы
        example: Дима
        type: string
      previousHozId:
        description: ID предыдущего хозяйства, когда корову продают, она переходит
          к новому владельцу и становится "новой коровой"
        type: integer
      rshnnumber:
        description: РСХН номер коровы
        example: "1323323232"
        type: string
      selecsNumber:
        description: Селекс номер коровы
        example: 98989
        type: integer
      sexId:
        description: ID пола коровы
        example: 1
        type: integer
    type: object
...
  models.DateOnly:
    properties:
      time.Time:
        type: string
    type: object
  models.Document:
    properties:
      cowID:
        description: ID коровы, для которой хранитя документ
        type: integer
      id:
        description: ID
        type: integer
      path:
        description: путь к документу относительно genmilk.ru/api/static/documents
        type: string
    type: object
...
  models.Exterior:
    properties:
      acrumLength:
        type: number
      bodyDepth:
        type: number
      bodyStructure:
        type: number
      boneQHockJointRear:
        type: number
      centerLigamentDepth:
        type: number
      chestWidth:
        type: number
      conditioning:
        type: number
      cowID:
        type: integer
      exteriorType:
        type: number
      foreLegPosFront:
        type: number
      foreTeatDiameter:
        type: number
      foreTeatLendth:
        type: number
      foreUdderAttach:
        type: number
      foreUdderPlcRear:
        type: number
      foreUdderWidth:
        type: number
      harmonyOfMovement:
        type: number
      heightOfUdderAttach:
        type: number
      hindLegPosRead:
        type: number
      hindLegPosSide:
        type: number
      hindTeatPlc:
        type: number
      hindUdderWidth:
        type: number
      hoofAngle:
        type: number
      id:
        type: integer
      limbs:
        type: number
      milkStrength:
        type: number
      pelvicWidth:
        type: number
      picturePath:
        type: string
      prominenceOfMilkVeins:
        type: number
      rating:
        type: number
      rearTeatDiameter:
        type: number
      rearTeatLength:
        type: number
      sacrumAngle:
        type: number
      sacrumHeight:
        type: number
      topLine:
        type: number
      udder:
        type: number
      udderBalance:
        type: number
      udderDepth:
        type: number
    type: object
  models.Farm:
    properties:
      address:
        description: Адрес
        type: string
      cowsCount:
        description: количество коров
        type: integer
      description:
        description: описание
        type: string
      districtId:
        description: ID района, в котором находится хозяйство
        type: integer
      email:
        description: эл. почта
        type: string
      hozNumber:
        description: |-
          Region   Region `json:"-"`
          RegionId uint
        type: string
      id:
        type: integer
      inn:
        description: ИНН
        type: string
      name:
        description: Название хозяйства
        type: string
      nameShort:
        description: Краткое название хозйства
        type: string
      parrentId:
        description: ID более управляющего хоз-ва (для хозяйства - холдинг, для фермы
          - хозяйство)
        type: integer
      phone:
        description: телефон
        type: string
      type:
        description: 'Тип: хозяйство, ферма, холдинг'
        type: integer
    type: object
  models.Genetic:
    properties:
      bloodDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата взятия пробы крови
      cowID:
        description: ID коровы
        type: integer
      geneticIllnessesData:
        description: Список генетических заболеваний, пустой если нет
        items:
          $ref: '#/definitions/models.GeneticIllnessData'
        type: array
      gtcFilePath:
        description: путь к gtc файлу относительно genmilk.ru/api/static/gtc
        type: string
      id:
        description: ID записи о генотипировании
        type: integer
      inbrindingCoeffByGenotype:
        description: Коэф. инбриндинга по генотипу
        type: number
      probeNumber:
        description: Номер пробы
        type: string
      resultDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата получения  результата
    type: object
  models.GeneticIllness:
    properties:
      description:
        description: описание генетического заболевания
        type: string
      id:
        type: integer
      name:
        description: имя генетического заболевания
        type: string
      omia:
        description: Какой-то там ОМИЯ номер
        type: string
    type: object
  models.GeneticIllnessData:
    properties:
      geneticID:
        description: ID данных о генотипировании коровы, к которым относятся данные
          о заболеваниях коровы
        type: integer
      id:
        type: integer
      illness:
        $ref: '#/definitions/models.GeneticIllness'
      illnessID:
        description: ID заболевание
        type: integer
      status:
        $ref: '#/definitions/models.GeneticIllnessStatus'
      statusID:
        description: ID статус заболевания
        type: integer
    type: object
  models.GeneticIllnessStatus:
    properties:
      id:
        type: integer
      status:
        description: 'Статус заболевания: FREE, CARIER, BAD ...'
        type: string
    type: object
...
host: genmilk.ru
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее
    Все даты передаются как строки
    Большая часть рутов не возвращает вложенные объекты
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GenMilk API
  version: "1.0"
paths:
...
  /auth/checkEmail:
    get:
      consumes:
      - application/json
      parameters:
      - description: email of user to check
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: {}
            type: object
      summary: CHECK EMAIL (USED FOR ADMIN PAGE)
      tags:
      - REGISTER
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Нужно для админки
      parameters:
      - description: applied filters
        in: body
        name: RegisterData
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: {}
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: {}
            type: object
      summary: REGISTER (FOR ADMIN PAGE)
      tags:
      - REGISTER
...
  /checkMilks:
    get:
      description: Возращает список контрольных доек. Без фильтра нельзя, т.к. слишком
        много контрольных доений
      parameters:
      - description: id лактации, для корой ищутся котнольные дойки
        in: query
        name: lactation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CheckMilk'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get list of checkMilks
      tags:
      - CheckMilks
  /checkMilks/{id}:
    get:
      description: Возращает контрольную дойку
      parameters:
      - description: id контрольной дойки
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CheckMilk'
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get checkMilk
      tags:
      - CheckMilks
...
  /cows/{id}/children:
    get:
      description: Возращает список всех детей для конкретной коровы.
      parameters:
      - description: ID коровы для которой ищутся лактации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cows.Children'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of children
      tags:
      - Cows
  /cows/{id}/documents:
    get:
      description: Возвращает список документов коровы
      parameters:
      - description: ID коровы для которой ищутся документы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of documents
      tags:
      - Cows
  /cows/{id}/exterior:
    get:
      description: Возращает информацию об экстерьере, null, если нет
      parameters:
      - description: ID коровы для которой ищется экстерьер
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exterior'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get exterior
      tags:
      - Cows
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"